refacto literal
