+++ Makefile
@@ -46,8 +46,10 @@ SRC = 	./src/main/main.c\
 		./src/utils/place_holder.c \
 		./src/utils/mini_split.c \
 		./src/lexer/lexer.c \
-		./src/lexer/redirection_check.c \
-		./src/lexer/literal.c
+		./src/lexer/literal/literal.c \
+		./src/lexer/literal/literal_string.c \
+		./src/lexer/literal/literal_redir.c \
+		./src/lexer/literal/literal_error.c
 
 # Colors #
 BLACK = \033[0;30m
