+++ src/lexer/interpreted/interpreted.c
@@ -6,7 +6,7 @@
 /*   By: ddemers <ddemers@student.42quebec.com>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/19 00:12:49 by ddemers           #+#    #+#             */
-/*   Updated: 2023/03/19 00:13:58 by ddemers          ###   ########.fr       */
+/*   Updated: 2023/03/19 01:14:21 by ddemers          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,5 +14,5 @@
 
 char	**interpreted_tokenization(t_mini *mini, char **literal)
 {
-	
+	return (NULL);
 }
